Move webiste_hasher, update references
