--- common/utils/__init__.py
@@ -1,2 +1,0 @@
-from .file_downloaders.downloaders import *
-from .website_hasher.page_update import hash_comparer, page_hasher, page_update
