+++ scrapers/data_portals/crimegraphics/crimegraphics_bulletin.py
@@ -11,9 +11,9 @@ from pathlib import Path
 
 p = Path(__file__).resolve().parents[3]
 sys.path.insert(1, str(p))
-
-from common.utils import hash_comparer, page_hasher, page_update
-from scrapers.data_portals.crimegraphics.utils import data_parser
+print(sys.path)
+from utils.website_hasher.page_update import hash_comparer, page_hasher, page_update
+from scrapers.data_portals.crimegraphics.data_parser import data_parser
 
 # this function is used for gathering time stats
 def function_timer(stats):
