+++ scrapers/data_portals/opendata/opendata_scraper_2.py
@@ -10,7 +10,7 @@ import time
 p = Path(__file__).resolve().parents[3]
 sys.path.insert(1, str(p))
 
-from common.utils import page_update
+from utils.website_hasher.page_update import page_update
 
 
 # save_url = [
