+++ utils/__init__.py
@@ -3,3 +3,6 @@ from .meta import data_truncator
 from .pdf import list_pdf_v2
 from .pdf import list_pdf_v3
 from .pdf import single_pdf_scraper
+from .website_hasher import hash_comparer
+from .website_hasher import page_hasher
+from .website_hasher import page_update
